+++ app/models/carto/helpers/active_record_compatibility.rb
@@ -6,4 +6,12 @@ module Carto::ActiveRecordCompatibility
   def new_record?
     return new?
   end
+
+  def save!
+    return save(raise_on_failure: true)
+  end
+
+  def attributes
+    values.with_indifferent_access
+  end
 end