+++ lib/assets/javascripts/new-dashboard/components/Subscriptions/ModalSubscription.vue
@@ -175,6 +175,10 @@
                 Check your subscriptions
               </Button>
           </router-link>
+
+        </div>
+        <div class="grid u-flex__justify--center u-mt--24">
+          <span v-if="error" class="error-msg">There was an issue with the subscription. Please contact CARTO support (support@carto.com)</span>
         </div>
       </div>
     </div>
@@ -214,7 +218,8 @@ export default {
     return {
       currentMode: null,
       licenseStatus: false,
-      loading: false
+      loading: false,
+      error: false
     };
   },
   computed: {
@@ -299,6 +304,7 @@ export default {
       window.dataLayer.push({ event: 'requestDataset' });
     },
     async subscribe () {
+      this.error = false;
       this.loading = true;
       if (
         await this.$store.dispatch('catalog/fetchSubscribe', {
@@ -309,6 +315,9 @@ export default {
       ) {
         await this.$store.dispatch('catalog/fetchSubscriptionsList');
         this.currentMode = 'subscribed';
+        this.loading = false;
+      } else {
+        this.error = true;
         this.loading = false;
       }
     },
@@ -322,11 +331,12 @@ export default {
         })
       ) {
         await this.$store.dispatch('catalog/fetchSubscriptionsList');
-        this.loading = false;
         this.closeModal();
       }
+      this.loading = false;
     },
     async request () {
+      this.error = false;
       this.loading = true;
       if (
         await this.$store.dispatch('catalog/fetchSubscribe', {
@@ -336,13 +346,18 @@ export default {
       ) {
         await this.$store.dispatch('catalog/fetchSubscriptionsList');
         this.currentMode = 'requested';
-        this.loading = false;
+      } else {
+        this.error = true;
       }
+      this.loading = false;
     }
   },
   watch: {
     mode () {
       this.currentMode = this.mode;
+    },
+    isOpen () {
+      this.error = false;
     }
   }
 };
@@ -412,6 +427,8 @@ export default {
 }
 
 .button {
+  height: 36px;
+
   &:not(.is-loading) {
     .loading {
       display: none;
@@ -422,4 +439,9 @@ export default {
   }
 }
 
+.error-msg {
+  font-size: 12px;
+  line-height: 20px;
+  color: $red--500;
+}
 </style>